Improve tuist registry output, fix resolving registry transitive dependencies
